feat(protocol-designer): timeline scrubber to the overview page
