+++ app/src/pages/Desktop/Protocols/ProtocolDetails/ProtocolTimeline.tsx
@@ -27,11 +27,7 @@ export function ProtocolTimeline(): JSX.Element {
 
   return storedProtocol != null && storedProtocol.mostRecentAnalysis != null ? (
     <Box padding={SPACING.spacing16}>
-      <ProtocolTimelineScrubber
-        commands={storedProtocol.mostRecentAnalysis.commands}
-        analysis={storedProtocol.mostRecentAnalysis}
-        robotType={storedProtocol.mostRecentAnalysis.robotType}
-      />
+      <ProtocolTimelineScrubber analysis={storedProtocol.mostRecentAnalysis} />
     </Box>
   ) : (
     <Icon size="8rem" name="ot-spinner" spin />
